Refactored code to allow data re-validation
